WAT-50 Refactored name of Artist Profile Summary component to Maker ..., also removed the summary list components
