+++ .idea/workspace.xml
@@ -4,10 +4,15 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="a371227e-1556-4281-bf9d-e84dc3a6719d" name="WAT-50 Content summaries" comment="Ran lint">
+    <list default="true" id="a371227e-1556-4281-bf9d-e84dc3a6719d" name="WAT-50 Content summaries" comment="WAT-50 Created Maker summary profile blocks">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/app/components/ArtistProfileSummary.stories.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/app/components/ArtistProfileSummary.stories.tsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/app/components/ArtistProfileSummary.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/app/components/ArtistProfileSummary.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/components/ArtistProfileSummary.stories.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/app/components/MakerProfileSummary.stories.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/components/ArtistProfileSummary.translations.ts" beforeDir="false" afterPath="$PROJECT_DIR$/app/components/MakerProfileSummary.translations.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/components/ArtistProfileSummary.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/app/components/MakerProfileSummary.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/components/ArtistsSummary.stories.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/app/components/ArtistsSummary.stories.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/components/ArtistsSummary.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/app/components/ArtistsSummary.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/locales/en.ts" beforeDir="false" afterPath="$PROJECT_DIR$/app/locales/en.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/locales/nl.ts" beforeDir="false" afterPath="$PROJECT_DIR$/app/locales/nl.ts" afterDir="false" />
     </list>
     <list id="7f955583-8571-46fe-bf73-823af2e4bb22" name="Changes" comment="Added box and line illustrations, and created the hand drawn box component" />
     <list id="beb23f38-d23a-47e7-ba80-27679dd64e74" name="WAT-54 CLI" comment="WAT-54 CLI" />
@@ -203,6 +208,7 @@
     "prettierjs.PrettierConfiguration.Package": "/Users/lody/dev/projects/watershed-website/node_modules/prettier",
     "run.code.analysis.last.selected.profile": "pProject Default",
     "settings.editor.selected.configurable": "tasks.servers",
+    "spentTimePerTaskStorage.store": "{}",
     "tasks.open.task.update.state.enabled": "true",
     "timeTracker.activityTracker": "null",
     "timeTracker.comment": "",
@@ -213,17 +219,17 @@
     "timeTracker.isPaused": "false",
     "timeTracker.isPostAfterCommitEnabled": "false",
     "timeTracker.isPostedScheduled": "true",
-    "timeTracker.isRunning": "true",
+    "timeTracker.isRunning": "false",
     "timeTracker.isScheduledEnabled": "false",
     "timeTracker.isWhenProjectClosedEnabled": "false",
     "timeTracker.issueId": "WAT-50",
     "timeTracker.issueIdReadable": "WAT-50",
-    "timeTracker.pausedTime": "1283226",
+    "timeTracker.pausedTime": "0",
     "timeTracker.query": "",
-    "timeTracker.recordedTime": "23",
+    "timeTracker.recordedTime": "90",
     "timeTracker.scheduledPeriod": "19:00:0",
-    "timeTracker.startTime": "29979710",
-    "timeTracker.timeInMills": "1417594",
+    "timeTracker.startTime": "36666207",
+    "timeTracker.timeInMills": "0",
     "timeTracker.type": "Development",
     "ts.external.directory.path": "/Users/lody/dev/projects/watershed-website/node_modules/typescript/lib",
     "vue.rearranger.settings.migration": "true"
@@ -774,7 +780,7 @@
       <option name="number" value="50" />
       <option name="presentableId" value="WAT-50" />
       <option name="project" value="WAT" />
-      <updated>1769881004040</updated>
+      <updated>1769888939370</updated>
       <workItem from="1760341706290" duration="17131000" />
       <workItem from="1760359568103" />
       <workItem from="1760434553239" duration="6105000" />
@@ -1022,7 +1028,6 @@
     </option>
   </component>
   <component name="VcsManagerConfiguration">
-    <MESSAGE value="We now have a basic workflow for editing the menu" />
     <MESSAGE value="Swapped navigation menu out for the actual header (in demo mode), which also meant we can load the dynamic header on the site itself. Also swapped orientation on the manage menu page" />
     <MESSAGE value="Extracted save action and implemented it on the page level, just like Cancel." />
     <MESSAGE value="Semantically fixed db migrations workflow" />
@@ -1047,7 +1052,8 @@
     <MESSAGE value="Refactor root props structure and add migration support for title visibility" />
     <MESSAGE value="Fixed lint errors and returned editor header in Puck overrides so that we could have our functionality back" />
     <MESSAGE value="Ran lint" />
-    <option name="LAST_COMMIT_MESSAGE" value="Ran lint" />
+    <MESSAGE value="WAT-50 Created Maker summary profile blocks" />
+    <option name="LAST_COMMIT_MESSAGE" value="WAT-50 Created Maker summary profile blocks" />
   </component>
   <component name="XSLT-Support.FileAssociations.UIState">
     <expand />
