--- app/components/ArtistsSummary.translations.ts
@@ -1,20 +1,0 @@
-export const en = {
-  ArtistsSummary: {
-    Title_one: 'About the creator',
-    Title_other: 'About the creators',
-  },
-};
-
-export const nl = {
-  ArtistsSummary: {
-    Title_one: 'Over de maker',
-    Title_other: 'Over de makers',
-  },
-};
-
-export const pap = {
-  ArtistsSummary: {
-    Title_one: 'About the creator',
-    Title_other: 'About the creators',
-  },
-};
